[semantic-arc] Split load [copy] -> load_borrow opt into its own file.
